add conformation before delete

+++ FlowHistoryCleaner/Program.cs
@@ -11,20 +11,28 @@ namespace FlowHistoryCleaner
     {
         static void Main(string[] args)
         {
-            
+
             string conn_str = Environment.GetEnvironmentVariable("AzureWebJobsStorage");
-            Console.WriteLine($"will use the conniction string {conn_str} ");
-            Console.ReadLine();
+            Console.WriteLine($"will use the connection string {conn_str} ");
+            if (!PromptConfirmation("Are you sure you want to clean all  Tables and Queues ?"))
+            {
+                return;
+            }
+
             var TserviceClient = new TableServiceClient(conn_str);
-         
+
             Pageable<TableItem> queryTableResults = TserviceClient.Query();
             foreach (TableItem table in queryTableResults)
             {
-                TserviceClient.DeleteTable(table.Name);
-                Console.WriteLine($"-Table {table.Name}");
+                if (table.Name.StartsWith("flow"))
+                {
+                    TserviceClient.DeleteTable(table.Name);
+                    Console.WriteLine($"-Table {table.Name} is Deleted ");
+                }
+
             }
 
-            
+
             QueueServiceClient QserviceClient = new QueueServiceClient(conn_str);
 
             //list all queues in the storage account
@@ -33,10 +41,14 @@ namespace FlowHistoryCleaner
             //then you can write code to list all the queue names          
             foreach (Azure.Page<QueueItem> queuePage in myqueues)
             {
-                foreach (QueueItem q in queuePage.Values)
+                foreach (QueueItem queue in queuePage.Values)
                 {
-                    QserviceClient.DeleteQueue(q.Name);
-                    Console.WriteLine($"-Queue {q.Name}" );
+                    if (queue.Name.StartsWith("flow"))
+                    {
+                        QserviceClient.DeleteQueue(queue.Name);
+                        Console.WriteLine($"-Queue {queue.Name} is Deleted ");
+                    }
+
 
                 }
 
@@ -44,5 +56,14 @@ namespace FlowHistoryCleaner
 
             Console.WriteLine("Done");
         }
+
+        static private bool PromptConfirmation(string confirmText)
+        {
+            Console.Write(confirmText + " [y/n] : ");
+            ConsoleKey response = Console.ReadKey(false).Key;
+            Console.WriteLine();
+            return (response == ConsoleKey.Y);
+        }
+
     }
 }
